@Consumer allows for prefix definition
- Move exploreMethodMetadata to metadata.accessor.ts
